feat(model): hashing order
