+++ apps/ai-dial-admin/src/components/Common/Multiselect/Modal/MultiselectModal.tsx
@@ -1,19 +1,16 @@
 import { FC, useCallback, useEffect, useRef, useState } from 'react';
-import { IconPlus } from '@tabler/icons-react';
 
 import Button from '@/src/components/Common/Button/Button';
-import Checkbox from '@/src/components/Common/Checkbox/Checkbox';
 import Loader from '@/src/components/Common/Loader/Loader';
 import Popup from '@/src/components/Common/Popup/Popup';
 import { ButtonsI18nKey } from '@/src/constants/i18n';
-import { BASE_ICON_PROPS } from '@/src/constants/main-layout';
 import { useNotification } from '@/src/context/NotificationContext';
 import { useI18n } from '@/src/locales/client';
+import { ServerActionResponse } from '@/src/models/server-action';
 import { PopUpState } from '@/src/types/pop-up';
 import { getErrorNotification } from '@/src/utils/notification';
-import { ServerActionResponse } from '@/src/models/server-action';
-import NewItemInput from './NewItemInput';
 import { uniq } from 'lodash';
+import MultiselectContentModal from './ModalContent';
 
 interface Props {
   initSelectedItems?: string[];
@@ -22,6 +19,8 @@ interface Props {
   addTitle?: string;
   addPlaceholder?: string;
   allItems?: string[];
+  draggable?: boolean;
+  editMode?: boolean;
   onClose: () => void;
   onSelectItems?: (items: string[]) => void;
   getItems?: () => Promise<ServerActionResponse>;
@@ -31,12 +30,11 @@ const MultiselectModal: FC<Props> = ({
   initSelectedItems,
   modalState,
   heading,
-  addTitle,
-  addPlaceholder,
   allItems,
   onClose,
   onSelectItems,
   getItems,
+  ...props
 }) => {
   const t = useI18n();
 
@@ -48,43 +46,11 @@ const MultiselectModal: FC<Props> = ({
   const [selectedItems, setSelectedItems] = useState<string[]>(initSelectedItems || []);
   const [items, setItems] = useState<string[]>([]);
   const [newItems, setNewItems] = useState<string[]>([]);
-  const newItemsContainer = useRef<HTMLDivElement | null>(null);
 
   const onApply = useCallback(() => {
     onSelectItems?.([...selectedItems, ...newItems].filter((t) => t !== ''));
     onClose();
   }, [onSelectItems, selectedItems, newItems, onClose]);
-
-  const onChangeSelectedItems = useCallback(
-    (topic: string, value?: boolean) => {
-      if (value) {
-        setSelectedItems([...selectedItems, topic]);
-      } else {
-        setSelectedItems(selectedItems.filter((t) => t !== topic));
-      }
-    },
-    [setSelectedItems, selectedItems],
-  );
-
-  const onChangeItem = useCallback(
-    (topic: string, index: number) => {
-      newItems[index] = topic.trimStart();
-      setNewItems([...newItems]);
-    },
-    [setNewItems, newItems],
-  );
-
-  const onRemoveItem = useCallback(
-    (index: number) => {
-      newItems.splice(index, 1);
-      setNewItems([...newItems]);
-    },
-    [setNewItems, newItems],
-  );
-
-  const onAddItem = useCallback(() => {
-    setNewItems([...newItems, '']);
-  }, [setNewItems, newItems]);
 
   useEffect(() => {
     const filtered = newItems.filter((v) => v !== '');
@@ -108,60 +74,19 @@ const MultiselectModal: FC<Props> = ({
     }
   }, [setItems, getItems, allItems]);
 
-  useEffect(() => {
-    const container = newItemsContainer.current;
-    if (container && container.scrollHeight > container.clientHeight) {
-      container.scrollTo({
-        top: container.scrollHeight,
-        behavior: 'smooth',
-      });
-    }
-  }, [newItems.length]);
-
   return (
     <Popup onClose={onClose} heading={heading} portalId="itemsMultiSelect" state={modalState}>
       <div className="flex flex-col overflow-auto px-6 py-4">
         {isLoading ? (
           <Loader size={40} />
         ) : (
-          <>
-            <div className="flex flex-col gap-y-2 overflow-auto" ref={newItemsContainer}>
-              {items.map((item, index) => {
-                return (
-                  <Checkbox
-                    key={index}
-                    checked={selectedItems.includes(item)}
-                    id={index.toString()}
-                    label={item}
-                    onChange={(v) => onChangeSelectedItems(item, v)}
-                  />
-                );
-              })}
-
-              {newItems.map((item, index) => {
-                return (
-                  <NewItemInput
-                    key={index}
-                    value={item}
-                    onChangeItem={onChangeItem}
-                    onRemoveItem={onRemoveItem}
-                    index={index}
-                    placeholder={addPlaceholder}
-                  />
-                );
-              })}
-            </div>
-            {addTitle && (
-              <div>
-                <Button
-                  cssClass="secondary mt-2"
-                  iconBefore={<IconPlus {...BASE_ICON_PROPS} />}
-                  title={addTitle}
-                  onClick={onAddItem}
-                />
-              </div>
-            )}
-          </>
+          <MultiselectContentModal
+            items={items}
+            selectedItems={selectedItems}
+            setSelectedItems={setSelectedItems}
+            setNewItems={setNewItems}
+            {...props}
+          />
         )}
       </div>
       <div className="flex flex-row items-center justify-end gap-2 px-6 py-4">
