+++ apps/ai-dial-admin/src/components/ModelView/ModelProperties/ModelProperties.tsx
@@ -64,6 +64,7 @@ const ModelProperties: FC<Props> = ({ model, modelsNames, updateModel }) => {
           <Multiselect
             elementId="order"
             draggable={true}
+            editMode={true}
             selectedItems={model.fieldsHashingOrder || []}
             allItems={model.fieldsHashingOrder || []}
             onChangeItems={(fieldsHashingOrder) => {
